+++ docs/make.jl
@@ -6,53 +6,15 @@ ENV["GKSwstype"] = "nul"
 
 bib = CitationBibliography(joinpath(@__DIR__, "bibliography.bib"))
 
-const EXAMPLES_DIR = joinpath(@__DIR__, "..", "examples")
-const OUTPUT_DIR = joinpath(@__DIR__, "src/literated")
-
-example_scripts = ["density_from_temperature_pressure_humidity.jl"]
-
-if isdir(OUTPUT_DIR)
-    rm(OUTPUT_DIR)
-end
-
-mkdir(OUTPUT_DIR)
-
-cp(
-    joinpath(EXAMPLES_DIR, "JRA55_atmospheric_state_Jan_1_1991.jld2"),
-    joinpath(OUTPUT_DIR, "JRA55_atmospheric_state_Jan_1_1991.jld2"),
-)
-
-for example in example_scripts
-    example_filepath = joinpath(EXAMPLES_DIR, example)
-
-    withenv("JULIA_DEBUG" => "Literate") do
-        start_time = time_ns()
-        Literate.markdown(
-            example_filepath,
-            OUTPUT_DIR;
-            flavor = Literate.DocumenterFlavor(),
-            execute = true,
-        )
-        elapsed = 1e-9 * (time_ns() - start_time)
-        @info @sprintf("%s example took %s seconds to build.", example, elapsed)
-    end
-end
-
-example_pages = [
-    "Density from temperature, pressure, and humidity" => "literated/density_from_temperature_pressure_humidity.md",
-]
-
 pages = Any[
     "Home" => "index.md",
-    "Installation" => "Installation.md",
-    "API" => "API.md",
-    "How-to-guide" => "HowToGuide.md",
-    "Examples" => example_pages,
-    "Tested profiles" => "TestedProfiles.md",
-    "Temperature profiles" => "TemperatureProfiles.md",
-    "Developer docs" => "DevDocs.md",
-    "Clausius Clapeyron relation" => "Clausius_Clapeyron.md",
-    "Thermodynamics overview" => "Formulation.md",
+    "Mathematical Formulation" => "Formulation.md",
+    "API Reference" => "API.md",
+    "How-To Guide" => "HowToGuide.md",
+    "Temperature Profiles" => "TemperatureProfiles.md",
+    "Tested Profiles" => "TestedProfiles.md",
+    "Clausius-Clapeyron Validation" => "Clausius_Clapeyron.md",
+    "Saturation Adjustment Convergence" => "SaturationAdjustmentConvergence.md",
     "References" => "References.md",
 ]
 
@@ -72,6 +34,7 @@ format = Documenter.HTML(
     mathengine = mathengine,
     collapselevel = 1,
     size_threshold = 2MiB,
+    assets = ["src/assets/custom.css"],
 )
 
 makedocs(;